requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.2.0
autopep8==2.3.1
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
ciso8601==2.3.1
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.4
fastapi-cli==0.0.5
fastjsonschema==2.20.0
frozenlist==1.5.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
h11==0.14.0
hikari==2.1.0
hikari-lightbulb==3.0.0a14
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
ipython==8.12.3
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
mongo==0.2.0
mongomock==4.2.0.post1
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
oauthlib==3.2.2
orjson==3.10.11
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pure_eval==0.2.3
pyaml==24.9.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pylint==3.3.1
pymongo==4.10.1
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
sentinels==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.2
tinycss2==1.4.0
tomlkit==0.13.2
tornado==6.4.1
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.1
yarg==0.1.9
yarl==1.17.1
